Move component IR tests to ui/components/ui/__tests__/ for colocation
- Move Button, Checkbox, Accordion tests from packages/test/__tests__/
  to ui/components/ui/__tests__/ (colocated with components)
- packages/test/__tests__/ now only has package API tests (error detection)
- Add @barefootjs/test to @barefootjs/ui devDependencies
- Add "bun" export condition to @barefootjs/test for dev-time src imports
