+++ ui/components/ui/__tests__/checkbox.test.ts
@@ -1,0 +1,68 @@
+import { describe, test, expect } from 'bun:test'
+import { readFileSync } from 'fs'
+import { resolve } from 'path'
+import { renderToTest } from '@barefootjs/test'
+
+const checkboxSource = readFileSync(resolve(__dirname, '../checkbox.tsx'), 'utf-8')
+
+describe('Checkbox', () => {
+  const result = renderToTest(checkboxSource, 'checkbox.tsx')
+
+  test('has no compiler errors', () => {
+    const realErrors = result.errors.filter(e => e.code !== 'BF043')
+    expect(realErrors).toEqual([])
+  })
+
+  test('isClient is true', () => {
+    expect(result.isClient).toBe(true)
+  })
+
+  test('componentName is Checkbox', () => {
+    expect(result.componentName).toBe('Checkbox')
+  })
+
+  test('has signals: internalChecked, controlledChecked', () => {
+    expect(result.signals).toContain('internalChecked')
+    expect(result.signals).toContain('controlledChecked')
+  })
+
+  test('renders as <button>', () => {
+    const button = result.find({ tag: 'button' })
+    expect(button).not.toBeNull()
+  })
+
+  test('has role=checkbox', () => {
+    const button = result.find({ role: 'checkbox' })
+    expect(button).not.toBeNull()
+    expect(button!.tag).toBe('button')
+  })
+
+  test('has aria-checked attribute', () => {
+    const button = result.find({ role: 'checkbox' })!
+    expect(button.aria).toHaveProperty('checked')
+  })
+
+  test('has data-state attribute', () => {
+    const button = result.find({ role: 'checkbox' })!
+    expect(button.dataState).not.toBeNull()
+  })
+
+  test('has click event handler', () => {
+    const button = result.find({ role: 'checkbox' })!
+    expect(button.events).toContain('click')
+  })
+
+  test('contains conditional SVG child (checkmark)', () => {
+    const svg = result.find({ tag: 'svg' })
+    expect(svg).not.toBeNull()
+  })
+
+  test('toStructure() includes role and aria info', () => {
+    const structure = result.toStructure()
+    expect(structure).toContain('[role=checkbox]')
+    expect(structure).toContain('[aria-checked]')
+    expect(structure).toContain('(click)')
+    expect(structure).toContain('svg')
+    expect(structure).toContain('path')
+  })
+})
